requirements.txt
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
Flask
gunicorn==20.0.4
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
requests==2.28.2
urllib3==1.26.14
Werkzeug==2.2.2
DateTime==5.1
flask-cors
ShopifyAPI
requests
Flask-Caching
gql[all]
